+++ src/probnum/filtsmooth/_gaussfiltsmooth/_gaussfiltsmooth_utils.py
@@ -1,0 +1,1 @@
+"""Utility functions for Gaussian filtering and smoothing."""
